+++ zellij-server/src/plugins/wasm_bridge.rs
@@ -2099,6 +2099,8 @@ fn check_event_permission(
         | Event::SoftKeyboardVisibilityChanged(..)
         | Event::HintText(..)
         | Event::ActivePaneScroll(..)
+        | Event::NestedSessionModeUpdate { .. }
+        | Event::NestedSessionKeybinds { .. }
         | Event::InputReceived => PermissionType::ReadApplicationState,
         Event::WebServerStatus(..) => PermissionType::StartWebServer,
         Event::PaneRenderReport(..) => PermissionType::ReadPaneContents,