+++ zellij-server/src/tab/mod.rs
@@ -4086,6 +4086,18 @@ impl Tab {
             pane.set_guest_session_name(session_name);
         }
     }
+    pub fn guest_session_name_on_pane(&self, pane_id: PaneId) -> Option<String> {
+        self.tiled_panes
+            .get_pane(pane_id)
+            .or_else(|| self.floating_panes.get_pane(pane_id))
+            .or_else(|| {
+                self.suppressed_panes
+                    .values()
+                    .find(|s_p| s_p.1.pid() == pane_id)
+                    .map(|s_p| &s_p.1)
+            })
+            .and_then(|pane| pane.guest_session_name())
+    }
     pub fn set_guest_modal_shortcuts_on_pane(
         &mut self,
         pane_id: PaneId,